// apps/web/src/api/model/apiHydratedBusiness.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { ApiHydratedBusinessStatus } from './apiHydratedBusinessStatus';

export interface ApiHydratedBusiness {
  /** Business name */
  name: string;
  /** Business email */
  email: string;
  /** Owner user ID */
  owner: string;
  /** Onboard survey ID */
  onboardSurvey?: string;
  /** Business KYC ID */
  kyc?: string;
  /** Business status */
  status: ApiHydratedBusinessStatus;
  /** Document id */
  id: string;
  /** Document created at */
  createdAt: string;
  /** Document updated at */
  updatedAt: string;
}