// apps/web/src/api/model/apiHydratedBusinessKYC.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { BusinessLeadership } from './businessLeadership';
import type { KYCBusinessAddress } from './kYCBusinessAddress';
import type { KYCBusinessInformation } from './kYCBusinessInformation';
import type { KYCDocumentation } from './kYCDocumentation';

export interface ApiHydratedBusinessKYC {
  /** Business ID */
  business: string;
  /** Business information */
  information?: KYCBusinessInformation;
  /** Business address */
  address?: KYCBusinessAddress;
  /** Business directors */
  directors?: BusinessLeadership[];
  /** KYC documentation */
  documentation?: KYCDocumentation;
  /** Whether KYC is completed */
  completed: boolean;
  /** Approval requested at */
  approvalRequestedAt?: string;
  /** Approved at */
  approvedAt?: string;
  /** Document id */
  id: string;
  /** Document created at */
  createdAt: string;
  /** Document updated at */
  updatedAt: string;
}