// apps/web/src/api/model/kYCBusinessInformationDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */

export interface KYCBusinessInformationDto {
  businessName: string;
  registrationType: string;
  registrationNumber: string;
  businessDescription: string;
  tin?: string;
}