// apps/web/src/api/model/kYCBusinessAddressDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */

export interface KYCBusinessAddressDto {
  /** Primary City */
  city: string;
  /** Primary Address Line One */
  state: string;
  /** Primary Country Code */
  countryCode: string;
  /** Phone Number with Calling Code */
  phone: string;
  /** Primary Address Line One */
  addressLineOne: string;
  /** Primary Address Line Two */
  addressLineTwo?: string;
  /** Email Address */
  email: string;
}